+++ src/Attributes/Description.php
@@ -4,7 +4,7 @@ namespace BenSampo\Enum\Attributes;
 
 use Attribute;
 
-#[Attribute(Attribute::TARGET_CLASS_CONSTANT)]
+#[Attribute(Attribute::TARGET_CLASS_CONSTANT | Attribute::TARGET_CLASS)]
 class Description
 {
     public function __construct(