+++ src/Enum.php
@@ -322,7 +322,7 @@ abstract class Enum implements EnumContract, Castable, Arrayable, JsonSerializab
         return
             static::getLocalizedDescription($value) ??
             static::getAttributeDescription($value) ??
-            static::getFriendlyKeyName(static::getKey($value));
+            static::getFriendlyName(static::getKey($value));
     }
 
     /**
@@ -355,17 +355,39 @@ abstract class Enum implements EnumContract, Castable, Arrayable, JsonSerializab
         if ($constReflection === false) {
             return null;
         }
+
         $descriptionAttributes = $constReflection->getAttributes(Description::class);
 
-        if (count($descriptionAttributes) === 1) {
-            return $descriptionAttributes[0]->newInstance()->description;
-        }
+        return match(count($descriptionAttributes)) {
+            0 => null,
+            1 => $descriptionAttributes[0]->newInstance()->description,
+            default => throw new Exception('You cannot use more than 1 description attribute on ' . class_basename(static::class) . '::' . $constantName),
+        };
+    }
 
-        if (count($descriptionAttributes) > 1) {
-            throw new Exception('You cannot use more than 1 description attribute on ' . class_basename(static::class) . '::' . $constantName);
-        }
+    /**
+     * Get the description of the enum class.
+     * Default to Enum class short name
+     *
+     * @return string
+     */
+    public static function getClassDescription(): string
+    {
+        return static::getClassAttributeDescription()
+            ?? static::getFriendlyName(self::getReflection()->getShortName());
+    }
 
-        return null;
+    protected static function getClassAttributeDescription(): ?string
+    {
+        $reflection = self::getReflection();
+
+        $descriptionAttributes = $reflection->getAttributes(Description::class);
+
+        return match (count($descriptionAttributes)) {
+            0 => null,
+            1 => $descriptionAttributes[0]->newInstance()->description,
+            default => throw new Exception('You cannot use more than 1 description attribute on '.class_basename(static::class))
+        };
     }
 
     /**
@@ -456,15 +478,15 @@ abstract class Enum implements EnumContract, Castable, Arrayable, JsonSerializab
     }
 
     /**
-     * Transform the key name into a friendly, formatted version.
+     * Transform the name into a friendly, formatted version.
      */
-    protected static function getFriendlyKeyName(string $key): string
+    protected static function getFriendlyName(string $name): string
     {
-        if (ctype_upper(preg_replace('/[^a-zA-Z]/', '', $key))) {
-            $key = strtolower($key);
+        if (ctype_upper(preg_replace('/[^a-zA-Z]/', '', $name))) {
+            $name = strtolower($name);
         }
 
-        return ucfirst(str_replace('_', ' ', Str::snake($key)));
+        return ucfirst(str_replace('_', ' ', Str::snake($name)));
     }
 
     /**